requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
requests==2.26.0
pylint==2.4.4
Jinja2==3.0.1